requirements.txt
flask
flask-login
gunicorn
tensorflow
numpy
pandas
joblib
pillow